% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcast.R
\name{covidcast_signal}
\alias{covidcast_signal}
\title{Produce a data frame for one signal.}
\usage{
covidcast_signal(
  data_source,
  signal,
  start_day = NULL,
  end_day = NULL,
  geo_type = c("county", "hrr", "msa", "dma", "state"),
  geo_value = "*"
)
}
\arguments{
\item{data_source}{String identifying the data source to query. See the
\href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}{signal documentation}
for a list of available sources and signals.}

\item{signal}{String identifying the signal from that source to query.}

\item{start_day}{Query data beginning on this date. Date object, or string in
the form YYYYMMDD. If \code{start_day} is \code{NULL}, defaults to first day data is
available for this signal.}

\item{end_day}{Query data up to this date, inclusive. Date object or string
in the form YYYYMMDD. If \code{end_day} is \code{NULL}, defaults to the most recent
day data is available for this signal.}

\item{geo_type}{The geography type for which to request this data, such as
\code{"county"} or \code{"state"}. Available types are described in the COVIDcast
signal documentation. Defaults to \code{"county"}.}

\item{geo_value}{Which geography to return. The default, \code{"*"}, fetches all
geographies. To fetch a specific geography, specify its ID as a string.}
}
\value{
Data frame with matching data. Contains \code{geo_value}, \code{time_value},
\code{direction}, \code{value}, \code{stderr}, and \code{sample_size} columns. \code{geo_value}
identifies the location, such as a state name or county FIPS code;
\code{time_value} contains Date objects. \code{value} is the signal quantity
requested and \code{stderr} its standard error if available. \code{sample_size}
indicates the sample size available in that geography on that day; sample
size may not be available. \code{direction} uses a local linear fit to estimate
whether the signal in this region is currently increasing or decreasing.
Consult the signal documentation for more details.
}
\description{
Obtains data for selected date ranges for all geographic regions of the
United States. Available data sources and signals are documented in the
\href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}{COVIDcast signal documentation}.
Most (but not all) data sources are available at the county level, but the
API can also return data aggregated to metropolitan statistical areas,
hospital referral regions, or states, as desired, by using the \code{geo_type}
argument. View \code{vignette("covidcastR")} for detailed example usage.
}
\details{
Downloading large amounts of data may be slow, so this function prints
messages for each day of data it downloads. To suppress these, use
\code{\link[base:suppressMessages]{base::suppressMessages()}}, as in \code{suppressMessages(covidcast_signal("fb-survey", ...))}.
}
\examples{
\dontrun{
## fetches all counties from 2020-05-10 to the most recent available data:
covidcast_signal("fb-survey", "raw_cli", start_day = "20200510")
## fetches all counties on just 2020-05-10 and no other days
covidcast_signal("fb-survey", "raw_cli", start_day = "20200510",
                 end_day = "20200510")
## fetch all states on 2020-05-10, 2020-05-11, 2020-05-12
covidcast_signal("fb-survey", "raw_cli", start_day = "20200510",
                 end_day = "20200512", geo_type = "state")
## fetch all available data for just Pennsylvania
covidcast_signal("fb-survey", "raw_cli", geo_type = "state",
                 geo_value = "pa")
## fetch all available data in Pittsburgh metropolitan area (identified by
## CBSA ID)
covidcast_signal("fb-survey", "raw_cli", geo_type = "msa",
                 geo_value = "38300")
}
}
\references{
COVIDcast API documentation:
\url{https://cmu-delphi.github.io/delphi-epidata/api/covidcast.html}

COVIDcast public map: \url{https://covidcast.cmu.edu/}
}
